Move check related factories to dedicated file
